move preproc_dataset function from example to openfmri.py in the main lib. openfmri_preproc.py keeps the command line interface
