+++ examples/openfmri_preproc.py
@@ -2,206 +2,77 @@
 
 """
 :Module: nipype_preproc_spm_openfmri
-:Synopsis: Preprocessing Openfmri
+:Synopsis: Command line interface to preprocess OpenfMRI
 :Author: yannick schwartz, dohmatob elvis dopgima
 
 """
 
-# standard imports
 import os
-import glob
-import warnings
+import sys
 
-# import spm preproc utilities
-from pypreprocess.nipype_preproc_spm_utils import (do_subjects_preproc,
-                                                   SubjectData)
-from pypreprocess.datasets import fetch_openfmri
+from optparse import OptionParser
 
-DATASET_DESCRIPTION = """\
-<p><a href="https://openfmri.org/data-sets">openfmri.org datasets</a>.</p>
-"""
+from pypreprocess.openfmri import preproc_dataset
 
 
-def preproc_dataset(data_dir, output_dir,
-                    ignore_subjects=None, restrict_subjects=None,
-                    delete_orient=False, dartel=False,
-                    n_jobs=-1):
-    """Main function for preprocessing a dataset with the OpenfMRI layout.
+parser = OptionParser(usage=(
+    '%prog [input_dir] [output_dir]\n\n'
+    'Examples:\n\r'
+    'openfmri_preproc.py /tmp/ds105 /tmp/ds105_preproc -s sub001 -O\n'
+    'openfmri_preproc.py /tmp/ds105 /tmp/ds105_preproc -O -D -n 6'))
 
-    Parameters
-    ----------
-    data_dir: str
-        Path of input directory. If does not exist and finishes
-        by a valid OpenfMRI dataset id, it will be downloaded,
-        i.e., /path/to/dir/{dataset_id}.
-    output_dir: str
-        Path of output directory.
-    ignore_subjects: list or None
-        List of subject identifiers not to process.
-    restrict_subjects: list or None
-        List of subject identifiers to process.
-    delete_orient: bool
-        Delete orientation information in nifti files.
-    dartel: bool
-        Use dartel.
-    n_jobs: int
-        Number of parallel jobs.
+parser.description = (
+    '`input_dir` is the path to an existing '
+    'OpenfMRI dataset or where to download it. '
+    'The directory name must match a valid OpenfMRI dataset id, '
+    'and therefore look like /path/to/dir/{dataset_id}. OpenfMRI datasets '
+    'identifiers may be found here: https://openfmri.org/data-sets but '
+    'contain only 3 digits. e.g., the valid id for ds000001 is ds001.')
 
-    Examples
-    --------
-    preproc_dataset('/tmp/ds105', '/tmp/ds105_preproc ',
-                    ignore_subjects=['sub002', 'sub003'],
-                    delete_orient=True,
-                    n_jobs=3)
+parser.add_option(
+    '-s', '--subjects', dest='subjects',
+    help=('Process a single subject matching the given id. '
+          'A file path may be given, and must contain '
+          'a subject_id per line.'))
 
-    Warning
-    -------
-    Subjects may be excluded if some data is missing.
+parser.add_option(
+    '-O', '--delete-orient', dest='delete_orient',
+    default=False, action="store_true",
+    help=('Delete orientation information in nifti files.'))
 
-    Returns list of Bunch objects with fields anat, func, and subject_id
-    for each preprocessed subject
-    """
-    parent_dir, dataset_id = os.path.split(data_dir)
+parser.add_option(
+    '-D', '--dartel', dest='dartel',
+    default=False, action="store_true",
+    help=('Use dartel.'))
 
-    if not os.path.exists(data_dir):
-        fetch_openfmri(parent_dir, dataset_id)
+parser.add_option(
+    '-n', '--n-jobs', dest='n_jobs', type='int',
+    default=os.environ.get('N_JOBS', '1'),
+    help='Number of parallel jobs.')
 
-    ignore_subjects = [] if ignore_subjects is None else ignore_subjects
+options, args = parser.parse_args(sys.argv)
+if len(args) < 3:
+    options, args = parser.parse_args(sys.argv + ['-h'])
+input_dir, output_dir = args[1:]
+input_dir = input_dir.rstrip('/')
+output_dir = output_dir.rstrip('/')
+_, dataset_id = os.path.split(input_dir)
 
-    # glob for subjects and their imaging sessions identifiers
-    if restrict_subjects is None:
-        subjects = [os.path.basename(x)
-                    for x in glob.glob(os.path.join(data_dir, 'sub???'))]
-    else:
-        subjects = restrict_subjects
+if not dataset_id.startswith('ds') and not os.path.exists(input_dir):
+    parser.error("The directory does not exist and "
+                 "does not seem to be an OpenfMRI dataset.")
 
-    subjects = sorted(subjects)
+if options.subjects is not None and os.path.exists(options.subjects):
+    with open(options.subjects, 'rb') as f:
+        restrict = f.read().split()
+else:
+    restrict = None if options.subjects is None else [options.subjects]
 
-    # producer subject data
-    def subject_factory():
-        for subject_id in subjects:
-            if subject_id in ignore_subjects:
-                continue
+preproc_dataset(data_dir=input_dir,
+                output_dir=output_dir,
+                restrict_subjects=restrict,
+                dartel=options.dartel,
+                delete_orient=options.delete_orient,
+                n_jobs=options.n_jobs)
 
-            sessions = set()
-            subject_dir = os.path.join(data_dir, subject_id)
-            for session_dir in glob.glob(os.path.join(
-                    subject_dir, 'BOLD', '*')):
-                sessions.add(os.path.split(session_dir)[1])
-            sessions = sorted(sessions)
-            # construct subject data structure
-            subject_data = SubjectData()
-            subject_data.session_id = sessions
-            subject_data.subject_id = subject_id
-            subject_data.func = []
-
-            # glob for BOLD data
-            has_bad_sessions = False
-            for session_id in subject_data.session_id:
-                bold_dir = os.path.join(
-                    data_dir, subject_id, 'BOLD', session_id)
-
-                # glob BOLD data for this session
-                func = glob.glob(os.path.join(bold_dir, "bold.nii.gz"))
-                # check that this session is OK (has BOLD data, etc.)
-                if not func:
-                    warnings.warn(
-                        'Subject %s is missing data for session %s.' % (
-                        subject_id, session_id))
-                    has_bad_sessions = True
-                    break
-
-                subject_data.func.append(func[0])
-
-            # exclude subject if necessary
-            if has_bad_sessions:
-                warnings.warn('Excluding subject %s' % subject_id)
-                continue
-
-            # anatomical data
-            subject_data.anat = os.path.join(
-                data_dir, subject_id, 'anatomy', 'highres001_brain.nii.gz')
-            if not os.path.exists(subject_data.anat):
-                subject_data.anat = os.path.join(
-                    data_dir, subject_id, 'anatomy', 'highres001.nii.gz')
-
-            # subject output_dir
-            subject_data.output_dir = os.path.join(output_dir, subject_id)
-            yield subject_data
-
-    return do_subjects_preproc(
-        subject_factory(),
-        n_jobs=n_jobs,
-        dataset_id=dataset_id,
-        output_dir=output_dir,
-        deleteorient=delete_orient,
-        dartel=dartel,
-        dataset_description=DATASET_DESCRIPTION,
-        # caching=False,
-        )
-
-if __name__ == '__main__':
-    import sys
-    from optparse import OptionParser
-
-    parser = OptionParser(usage=(
-        '%prog [input_dir] [output_dir]\n\n'
-        'Examples:\n\r'
-        'openfmri_preproc.py /tmp/ds105 /tmp/ds105_preproc -s sub001 -O\n'
-        'openfmri_preproc.py /tmp/ds105 /tmp/ds105_preproc -O -D -n 6'))
-
-    parser.description = (
-        '`input_dir` is the path to an existing '
-        'OpenfMRI dataset or where to download it. '
-        'The directory name must match a valid OpenfMRI dataset id, '
-        'and therefore look like /path/to/dir/{dataset_id}. OpenfMRI datasets '
-        'identifiers may be found here: https://openfmri.org/data-sets but '
-        'contain only 3 digits. e.g., the valid id for ds000001 is ds001.')
-
-    parser.add_option(
-        '-s', '--subjects', dest='subjects',
-        help=('Process a single subject matching the given id. '
-              'A file path may be given, and must contain '
-              'a subject_id per line.'))
-
-    parser.add_option(
-        '-O', '--delete-orient', dest='delete_orient',
-        default=False, action="store_true",
-        help=('Delete orientation information in nifti files.'))
-
-    parser.add_option(
-        '-D', '--dartel', dest='dartel',
-        default=False, action="store_true",
-        help=('Use dartel.'))
-
-    parser.add_option(
-        '-n', '--n-jobs', dest='n_jobs', type='int',
-        default=os.environ.get('N_JOBS', '1'),
-        help='Number of parallel jobs.')
-
-    options, args = parser.parse_args(sys.argv)
-    if len(args) < 3:
-        options, args = parser.parse_args(sys.argv + ['-h'])
-    input_dir, output_dir = args[1:]
-    input_dir = input_dir.rstrip('/')
-    output_dir = output_dir.rstrip('/')
-    _, dataset_id = os.path.split(input_dir)
-
-    if not dataset_id.startswith('ds') and not os.path.exists(input_dir):
-        parser.error("The directory does not exist and "
-                     "does not seem to be an OpenfMRI dataset.")
-
-    if options.subjects is not None and os.path.exists(options.subjects):
-        with open(options.subjects, 'rb') as f:
-            restrict = f.read().split()
-    else:
-        restrict = None if options.subjects is None else [options.subjects]
-
-    preproc_dataset(data_dir=input_dir,
-         output_dir=output_dir,
-         restrict_subjects=restrict,
-         dartel=options.dartel,
-         delete_orient=options.delete_orient,
-         n_jobs=options.n_jobs)
-
-    print "\r\nAll output written to %s" % output_dir
+print "\r\nAll output written to %s" % output_dir
